fix comments

+++ app.js
@@ -25,28 +25,42 @@ app.get("/allShop", (req, res) => {
 });
 
 app.post("/createShop", (req, res) => {
-  const buy = new Buy(req.body);
-  buy.save().then((result) => {
-    res.send({ data: result });
-  });
+  if (req.body.hasOwnProperty("shop") && req.body.hasOwnProperty("price") && req.body.hasOwnProperty("date")) {
+    const buy = new Buy(req.body);
+    buy.save().then((result) => {
+      res.send({ data: result });
+    });
+  } else {
+    res.status(404).send("Error");
+  }
 });
 
 app.delete("/deleteShop", (req, res) => {
-  Buy.deleteOne({ _id: req.query._id }).then((result) => {
-    Buy.find().then((result) => {
-      res.send({ data: result });
+  if (req.query._id) {
+    Buy.deleteOne({ _id: req.query._id }).then((result) => {
+      Buy.find().then((result) => {
+        res.send({ data: result });
+      });
     });
-  });
+  } else {
+    res.status(404).send("Error");
+  }
 });
 
 app.patch("/updateShop", (req, res) => {
-  console.log("req.body._id", req.body);
-  Buy.updateOne({ _id: req.body._id }, req.body).then((result) => {
-    // console.log('result', result);
-    Buy.find().then((result) => {
-      res.send({ data: result });
-    });
-  });
+  if (req.query._id){
+    if (req.body.hasOwnProperty("shop") || req.body.hasOwnProperty("price") || req.body.hasOwnProperty("date")) {
+      Buy.updateOne({ _id: req.body._id }, req.body).then((result) => {
+        Buy.find().then((result) => {
+          res.send({ data: result });
+        });
+      });
+    }else {
+      res.status(404).send("Error");
+    }
+  } else {
+    res.status(404).send("Error");
+  }
 });
 
 app.listen(8000, () => {
